+++ src/PommaLabs.KVLite.Database/DbCache.cs
@@ -29,7 +29,7 @@ using PommaLabs.KVLite.Goodies;
 using PommaLabs.KVLite.Resources;
 using PommaLabs.Thrower;
 using PommaLabs.Thrower.Goodies;
-using PommaLabs.Thrower.Logging;
+using PommaLabs.KVLite.Logging;
 using System;
 using System.Collections.Generic;
 using System.Data.Common;