+++ src/PommaLabs.KVLite/AbstractAsyncCache.cs
@@ -24,7 +24,7 @@
 using NodaTime;
 using PommaLabs.KVLite.Resources;
 using PommaLabs.Thrower;
-using PommaLabs.Thrower.Logging;
+using PommaLabs.KVLite.Logging;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;